refactor(tests): split by call kind
Accomodate for upcoming Enarx-specific call tests
