+++ tests/mod.rs
@@ -1,0 +1,1 @@
+mod integration_tests;
